Break-up solution graph tests
